refactor: move validation to utils
